# Reject invalid tags