--- a/docker/docker-cluster/metrictank.ini
+++ b/docker/docker-cluster/metrictank.ini
@@ -237,6 +237,10 @@ speculation-threshold = 1
 
 ## metric data inputs ##
 
+[input]
+# reject received metrics that have invalid tags
+reject-invalid-tags = true
+
 ### carbon input (optional)
 [carbon-in]
 enabled = true