--- a/docs/config.md
+++ b/docs/config.md
@@ -286,6 +286,13 @@ speculation-threshold = 1
 ```
 
 ## metric data inputs ##
+
+```
+[input]
+# reject received metrics that have invalid tags
+reject-invalid-tags = true
+```
+
 ### carbon input (optional)
 
 ```