--- a/idx/memory/memory.go
+++ b/idx/memory/memory.go
@@ -276,8 +276,10 @@ func (m *UnpartitionedMemoryIdx) Init() error {
 }
 
 func (m *UnpartitionedMemoryIdx) Stop() {
-	m.findCache.Shutdown()
-	m.findCache = nil
+	if m.findCache != nil {
+		m.findCache.Shutdown()
+		m.findCache = nil
+	}
 	return
 }