--- a/metrictank-sample.ini
+++ b/metrictank-sample.ini
@@ -240,6 +240,10 @@ speculation-threshold = 1
 
 ## metric data inputs ##
 
+[input]
+# reject received metrics that have invalid tags
+reject-invalid-tags = true
+
 ### carbon input (optional)
 [carbon-in]
 enabled = false