… vendor/github.com/raintank/schema/archive.go
… vendor/github.com/raintank/schema/archive_gen.go
… vendor/github.com/raintank/schema/archive_gen_test.go
… vendor/github.com/raintank/schema/key_gen_test.go
… vendor/github.com/raintank/schema/metric.go