This gives users the option to reject ingested metrics with invalid tags. Even when rejection is disabled it counts the metrics with invalid tags, this should make it easy for users to predict whether enabling the rejection is going to actually reject some of their metrics or not.
Also updates the vendored `raintank/schema` to the latest version, because it uses the new tag validation functions.

This is part of: https://github.com/grafana/grafana/issues/15261
As discussed in: https://github.com/grafana/metrictank/pull/1217#issuecomment-473824218
Depends on this PR to be merged, because otherwise with the latest `raintank/schema` the serialization of `cwr.ChunkWriteRequest` is broken: https://github.com/grafana/metrictank/pull/1335

The benchmarks look about the same with and without this change:
```
benchmark                                     old ns/op     new ns/op     delta
BenchmarkProcessMetricDataUniqueMetrics-3     656           731           +11.43%
BenchmarkProcessMetricDataSameMetric-3        657           754           +14.76%
benchmark                                     old allocs     new allocs     delta
BenchmarkProcessMetricDataUniqueMetrics-3     0              0              +0.00%
BenchmarkProcessMetricDataSameMetric-3        0              0              +0.00%
benchmark                                     old bytes     new bytes     delta
BenchmarkProcessMetricDataUniqueMetrics-3     5             5             +0.00%
BenchmarkProcessMetricDataSameMetric-3        5             5             +0.00%
```