@Dieterbe 2019-06-14
why is it optional? what happens to invalid tags if we "accept" them? and why would we accept them?