@replay 2019-06-14
It is optional because when enforced then it introduces new restrictions which previously have not been applied. That way some current user who isn't sure whether some of their metrics would get rejected can just disable the enforcement and look at the new `invalid_tags` metric to see if it increases or not, if it doesn't increase then none of their current metrics would get rejected.