@replay 2019-06-14
If invalid tags get accepted, as long as they still contain a `=` which is required for the index to be able to handle them, they will just get added to the index like others. If one does not contain a `=` at all, then the "invalid metric counter" gets increased and this tag gets ignore, but the metrics still gets added to the index via it's other tags and the name. If such invalid tags got accepted, it's then possible that they're not properly queriable, because graphite might not pass the query on.