@replay 2019-06-18
Now that https://github.com/grafana/metrictank/pull/1335 is merged I've rebased this onto master