@Dieterbe 2019-06-19
see https://github.com/grafana/metrictank/blob/master/docs/CONTRIBUTING.md#when-contributing-prs rule 16: potentially breaking change requires changelog update